pyproject.toml
[tool.poetry]
name = "events-app"
version = "0.1.0"
description = "Django REST API for event management"
authors = ["Alex Revuka <a.revuka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.16.0"
django = "^5.2.1"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"
drf-spectacular = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
